+++ bench/zbench.ml
@@ -1,5 +1,5 @@
 module Prim = struct
-  type t = Zero | One | Two | Three | Four
+  type t = Zero | Four
 
   let compare (x : t) (y : t) = Stdlib.compare x y
 
@@ -9,29 +9,15 @@ module Prim = struct
 
   let pp fmtr = function
     | Zero -> Format.fprintf fmtr "zero"
-    | One -> Format.fprintf fmtr "one"
-    | Two -> Format.fprintf fmtr "two"
-    | Three -> Format.fprintf fmtr "three"
     | Four -> Format.fprintf fmtr "four"
 
-  let arity = function
-    | Zero -> 0
-    | One -> 1
-    | Two -> 2
-    | Three -> 3
-    | Four -> 4
+  let arity = function Zero -> 0 | Four -> 4
 end
 
 module Pack = Term_indexing.Make (Prim)
 open Pack
 
 let zero = Term.prim Zero [||]
-
-let one t = Term.prim One [| t |]
-
-let two t0 t1 = Term.prim Two [| t0; t1 |]
-
-let three t0 t1 t2 = Term.prim Three [| t0; t1; t2 |]
 
 let four t0 t1 t2 t3 = Term.prim Four [| t0; t1; t2; t3 |]
 
@@ -46,7 +32,7 @@ let timeit f =
 
 (* ---------------------------------------------------------------- *)
 
-let path : Term.t -> Path.t =
+let rev_path : Term.t -> int list =
  fun t ->
   let rec aux path t =
     Term.destruct
@@ -55,11 +41,11 @@ let path : Term.t -> Path.t =
         if arity = 0 then path
         else
           let i = Random.int arity in
-          aux (Path.at_index i path) subterms.(i))
+          aux (i :: path) subterms.(i))
       (fun _ -> path)
       t
   in
-  aux Path.root t
+  aux [] t
 
 let rec large_term n =
   if n = 0 then zero
@@ -78,8 +64,8 @@ let rec guide_zip path zip =
 let () =
   let t = large_term 1_000_000 in
   let acc = ref 0.0 in
-  for i = 0 to 100 do
-    let p = path t |> Path.reverse in
+  for _ = 0 to 100 do
+    let p = rev_path t |> List.rev in
     let dt =
       timeit (fun () ->
           let z = guide_zip p (Zipper.of_term t) in